Upgrade to Java 8; code cleanups

+++ common/src/main/java/ch/algotrader/util/RoundUtil.java
@@ -79,7 +79,7 @@ public class RoundUtil {
         if (value == null) {
             return null;
         } else {
-            return getBigDecimal(value.doubleValue());
+            return getBigDecimal(value);
         }
     }
